remove redundant ontology metadata file
Motivation:

PaNET currently contains two files with the same information.  They
differ only in which format that information is expressed.  Tools exist
that allow for automatic translation of this information.  Moreover,
duplicating information violates the normalisation principle and risks
that inconsistencies are introduced.

Modification:

Remove the RDF/XML formatted file.  This leaves the Turtle version as
the definitive source of ontology metadata.

Result:

Duplicate information is removed.

Closes: ExPaNDS-eu/ExPaNDS-experimental-techniques-ontology#68