Create new GitHub workflow to autobuild PaNET
Motivation:

To avoid human processing (and corresponding risk of errors) we want to
automate as much as possible.

The `panet-build` container image provides the desired level of
automation, creating the PaNET artefacts and corresponding
documentation.

We would like to run this `panet-build` container image as a GitHub
workflow.

Modification:

Create entirely new GitHub workflow that runs the targeted build image.

The generated artefact (currently just `PaNET.owl`) is uploaded as an
artifact.

The WIDOCO-generated documentation is also uploaded as an artifact.

If the trigger is a new version tag then the webpages are updated using
the WIDOCO artifact.

Result:

PaNET is now build automatically.  The web pages are also updated
automatically when a new version is tagged.