Create new GitHub Workflow to automate certain tasks
Motivation:

To avoid human processing (and corresponding risk of errors) we want to
automate as much as possible.  This patch is the first step in this
direction.

The `panet-build` container image provides a reasonable level of
automation.  It will create the PaNET artefacts (currently limited to
the `PaNET.owl` file) and a set of web pages that provide documentation.

We would like to run this `panet-build` container image as a GitHub
workflow.  This should happen under two circumstances: when an update is
pushed to a branch and when a new version is tagged.

Modification:

Create a new GitHub workflow that runs the `panet-build` build image and
handles the resulting artefacts.

The generated artefacts (currently just `PaNET.owl`) are uploaded as a
GitHub artifact `PaNET`.  The WIDOCO-generated documentation is also
uploaded as the GitHub artifact `github-pages`.  These GitHub artifacts
are kept for seven days.

If the trigger is a new version tag then an additional job is executed
that takes the `github-pages` GitHub artifact and deploys the contents
as the new webpage.

Note: the workflow does NOT create a new release when a suitable tag is
created.  This is still a manual step.

Result:

PaNET is now build automatically, with the artefacts made available for
seven days.  The web pages are also updated automatically when a new
version is tagged.