Remove version information from PaNET metadata
Motivation:

The version information in `PaNET_metadata.ttl` is out-of-date.
Moreover the current concept is to use git tags to identify the version,
which makes this information redundant.

The PaNET build script removes the `owl:versionInfo` statements in
`PaNET_metadata.ttl` before adding the git-derived version information.
Although this results in correct version information appearing in the
PaNET file, the version information in `PaNET_metadata.ttl` is
potentially confusing.

Modification:

Remove `owl:versionInfo` from `PaNET_metadata.ttl`.

Result:

The build process is simpler to understand, as there are no longer any
statements in `PaNET_metadata.ttl` that are not copied into the final
ontology.

Closes: #132