Generate reasoned version of PaNET
Motivation:

From issue #120:
> [Support] users that want to exploit it using SPARQL queries, or
> simple python scripts. In the case of a clean Tbox the user looses all
> the inferred axioms and possibly some valuable results.

Modification:

Update build script to v0.5.  This includes support for building an
additional "reasoned" version of the ontology as
`source/PaNET_reasoned.owl`.

Update the workflow to capture this additional artefact when uploading
to GitHub.

Result:

We now build (and make available) both an unreasoned and a reasoned
version of the ontology.

Closes: #120