Partially solving Issue #150: spin echo scattering
Motivation

In order to improve the spelling, I suggest to delete the additional space in 'spin echo  scattering', PaNET01119.

Modification

Deleting this space directly affects the label of PaNET01119. Additionally this requires changing the spelling when this technique is used as Parent IRI in PaNET01120, PaNET01126, and PaNET01127.

Result

The ontology does not contain double spaces anymore. The test with panet-build was successful. The test with panet-build was successful.