Partially solving Issue #150: high resolution neutron imaging
Motivation

In order to unify the spelling, I suggest to be consistent in the use of dash or space when using 'high resolution'. Only the term 'high resolution neutron imaging' (PaNET01250) is using a dash currently.

Modification

Exchanging the dash with a space in 'high resolution neutron imaging' (PaNET01250). No further terms are affected by that change

Result

The notation of 'high resolution' is now unique within the ontology.