Partially solving Issue #150: infrared
Motivation

In order to unify the spelling, I suggest to be consistent in writing 'infrared'/'IR'. Only the term 'IR photon probe' (PaNET01005) solely contains the abbreviation, but does not have a (Alt) Label that is written in full.

Modification

'infrared photon probe' is now the Label of PaNET01005, with the Alt Label 'IR photon probe'. This induces changes of the Parent IRI in PaNET01109, PaNET01175, and PaNET01778.

Result

The notation of 'infrared' is consistent within the ontology. The test with panet-build was successful.