Partially solving Issue #150: time of flight
Motivation

In order to unify the spelling, I suggest to be consistent in writing 'TOF'/'time-of-flight'/'time of flight'. Only the term 'time-of-flight spectrometry' (PaNET1244) uses dashes. Only the term 'pulsed neutron TOF probe' (PaNET01019) solely uses 'TOF' without having an Alt Label with the fully written term.

Modification

The Label of PaNET1244 is now 'time of flight spectrometry'. The Label of PaNET01019 is now 'pulsed neutron time of flight probe' with the Alt Label 'pulsed neutron TOF probe'. No further changes are induced.

Result

The notation of 'time of flight' is consistent within the ontology. The test with panet-build was successful.