Add mapping of PaNET and voc4cat
Motivation

Mappings allow to connect different vocabularies or listed terms with each
other. voc4cat is an outcome of a NFDI initiative
(https://github.com/nfdi4cat/voc4cat-tool). voc4cat is a vocabulary of
catalytic terms that overlap with PaNET terms. Connecting the techniques
of voc4cat with PaNET terms will enable different applications in the future.

Modification

I used the skos terms relatedMatch, closeMatch, exactMatch, narrowMatch, and broadMatch to connect a PaNET term (IRI and name) to a term in WFL.the mapping between PaNET terms and related terms in WFL. I used the skos terms relatedMatch, closeMatch, exactMatch, narrowMatch, and broadMatch to connect a PaNET term (IRI and name) to a term in WFL.the mapping between PaNET terms and related terms in WFL. I used the skos terms relatedMatch, closeMatch, exactMatch, narrowMatch, and broadMatch to connect a PaNET term (IRI and name) to a term in WFL.the mapping between PaNET terms and related terms in WFL. I used the skos terms relatedMatch, closeMatch, exactMatch, narrowMatch, and broadMatch to connect a PaNET term (IRI and name) to a term in WFL.the mapping between PaNET terms and related terms in WFL. I used the skos terms relatedMatch, closeMatch, exactMatch, narrowMatch, and broadMatch to connect a PaNET term (IRI and name) to a term in WFL.t, I used the
skos terms relatedMatch, closeMatch, exactMatch, narrowMatch, and broadMatch
to connect a PaNET term (IRI and name) to a term in voc4cat.

Result

Future applications can now connect PaNET terms to the functionalities and
database content of WFL.

Closes #191.