Update label
Remove the 'technique' and 'concept' suffixes. In the event of label clashes, the IRI is used instead.

Instead of using the full IRI, CURIEs are used within axioms together with the prefix "PaNET:". CURIEs are an abbreviation for strings which are intended to represent IRIs.
https://www.w3.org/TR/curie/

For example, the IRI 'http://purl.org/pan-science/PaNET/PaNET00105' can be replaced by the CURIE 'PaNET:PaNET00105'. This is to make the axioms more compact.

The ROBOT diff function confirms that these changes do not affect the taxonomy.

Part of issue #250, explained in https://github.com/ExPaNDS-eu/ExPaNDS-experimental-techniques-ontology/issues/250#issuecomment-3275407811