Unifying the use of time of flight
Motivation

The term 'time of flight' is currently also used as 'time-of-flight' and 'TOF'.

Modification

The use of 'time-of-flight' was entirely replaced by 'time of flight'. An altlable including 'time of flight' was added, when the label only contained the abbreviation 'TOF'.

Result

Unified use of 'time of flight' within PaNET.

Contributes to #150 and #267, replaces #154.