Improving name and superclasses of micro scale diffraction
Motivation

The technique "micro scale diffraction" is an unusual term. Google and google scholar searches for different variations of this term resulted in one publication using it for a technique (not gratings etc). The publication references articles on "microbeam diffraction" and "microdiffraction" (with x-rays). However, the terms "microfocus (x-ray) diffraction" are missing in PaNET.

Modification

The label of "micro scale diffraction" is changed to "microfocus diffraction"; the original name is kept as altLabel. A definition was added. The superclass "microfocused probe" was added.

Result

The technique follows now the unwritten conventions of PaNET.