Adding definition and definition source for polymer micro- and nanografting
Motivation

Ther term "polymer micro- and nanografting" currently does not have a definition or definition source. It also only has one superclass: "manufacturing". There is not context given, what "micro" and "nano" refer to for this technique.

Modification

A literature reference for this technique was added. Further, this book was used to create and a short definition.

Result

Based on this reference, it is clear that "micro" and "nano" refer to the dimensions of the produced structure. Therefore, we currently do not have any useful specifiers that we could add as superclasses.

Contributes to #229