Update documentation webpage: no javascript
Motivation

The initial version of the documentation page was using javascript that might be harder to maintain than a
plain html page.

Modification

The page has been restructured such that the javascript file is not needed anymore.
The order in the navigation bar was updated and now shows the most general topics first, going to most specific.
The scope of PaNET was updated to give a short definition of PaNET and its idea, before going into too much detail.
A link has been added to the widoco page to browse PaNET.
A favicon has been created and added.

Result

A more user friendly page has been created.