Webpage for listing applications for and catalogs with PaNET
Motivation

We want to make PaNET as widely known as possible and demonstrate possible applications to new users.

Modification

A webpage was added that lists different catalogues in which PaNET is registered (might be incomplete). Additionally, the page also links existing services based on PaNET.

Result

Demonstrating relevance of PaNET and adding context.

Contributes to #148, closes #77